pyproject.toml
[tool.poetry]
name = "telegram-vpn"
version = "0.1.0"
description = ""
authors = ["Andrey Silnykh"]
license = "MIT"
readme = "README.md"
packages = [{include = "telegram_vpn"}]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.25.1"
envparse = "^0.2.0"
python-dotenv = "^1.0.0"
aiohttp = "^3.8.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
